Change ec2MetadataClient
