+++ src/shared/clients/ec2MetadataClient.ts
@@ -3,10 +3,37 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+import { MetadataService } from 'aws-sdk'
+import { ClassToInterface } from '../utilities/tsUtils'
 export interface InstanceIdentity {
     region: string
 }
 
-export interface Ec2MetadataClient {
-    getInstanceIdentity(): Promise<InstanceIdentity>
+export type Ec2MetadataClient = ClassToInterface<DefaultEc2MetadataClient>
+export class DefaultEc2MetadataClient implements Ec2MetadataClient {
+    private static readonly METADATA_SERVICE_TIMEOUT: number = 500
+
+    public constructor(private metadata: MetadataService = DefaultEc2MetadataClient.getMetadataService()) {}
+
+    getInstanceIdentity(): Promise<InstanceIdentity> {
+        return new Promise((resolve, reject) => {
+            this.metadata.request('/latest/dynamic/instance-identity/document', (err, response) => {
+                if (err) {
+                    reject(err)
+                }
+                const document: InstanceIdentity = JSON.parse(response)
+                resolve(document)
+            })
+        })
+    }
+
+    private static getMetadataService() {
+        return new MetadataService({
+            httpOptions: {
+                timeout: DefaultEc2MetadataClient.METADATA_SERVICE_TIMEOUT,
+                connectTimeout: DefaultEc2MetadataClient.METADATA_SERVICE_TIMEOUT,
+            } as any,
+            // workaround for known bug: https://github.com/aws/aws-sdk-js/issues/3029
+        })
+    }
 }
