+++ src/test/shared/extensionUtilities.test.ts
@@ -9,7 +9,7 @@ import { AWSError } from 'aws-sdk'
 import { writeFile, remove } from 'fs-extra'
 import * as path from 'path'
 import * as sinon from 'sinon'
-import { DefaultEc2MetadataClient } from '../../shared/clients/defaultEc2MetadataClient'
+import { DefaultEc2MetadataClient } from '../../shared/clients/ec2MetadataClient'
 import * as vscode from 'vscode'
 import {
     getComputeRegion,
